use product controller and stub out request handlers
